requirements.txt
Django
djangorestframework
dj-database-url

psycopg2
requests